requirements.txt
tweety-ns
discord-webhook
